Add `-Wsign-conversion -Wsign-promo` compiler flags (#43)
* Add `-Wsign-conversion -Wsign-promo` compiler flags

* chore: Update to Visual Studio 22 to match GitHub windows-latest image

+++ include/cpp-statsd-client/UDPSender.hpp
@@ -320,8 +320,8 @@ inline void UDPSender::sendToDaemon(const std::string& message) noexcept {
                             (struct sockaddr*)&m_server,
                             sizeof(m_server));
     if (ret == -1) {
-        m_errorMessage = "sendto server failed: host=" + m_host + ":" + std::to_string(m_port) +
-                         ", err=" + std::to_string(SOCKET_ERRNO);
+        m_errorMessage = "sendto server failed: host=" + m_host + ":" +
+                         std::to_string(static_cast<unsigned int>(m_port)) + ", err=" + std::to_string(SOCKET_ERRNO);
     }
 }
